refactor: split stats trends and streak helpers
